pub-sub-stream/requirements.txt
confluent-kafka==2.3.0
mysqlclient==2.2.1
